+++ Package.swift
@@ -47,6 +47,10 @@ let package = Package(
         .library(
             name: "StripeConnect",
             targets: ["StripeConnect"]
+        ),
+        .library(
+            name: "StripeCryptoOnramp",
+            targets: ["StripeCryptoOnramp"]
         )
     ],
     targets: [
@@ -173,6 +177,15 @@ let package = Package(
             name: "StripeConnect",
             dependencies: ["StripeCore", "StripeUICore", "StripeFinancialConnections"],
             path: "StripeConnect/StripeConnect"
+        ),
+        .target(
+            name: "StripeCryptoOnramp",
+            dependencies: ["StripeApplePay", "StripeCore", "StripeIdentity", "StripePayments", "StripePaymentSheet", "StripePaymentsUI", "StripeUICore"],
+            path: "StripeCryptoOnramp/StripeCryptoOnramp",
+            exclude: ["StripeCryptoOnramp.h"],
+            resources: [
+                .process("Resources/StripeCryptoOnramp.xcassets")
+            ]
         )
     ]
 )