+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Link/LinkController.swift
@@ -360,15 +360,15 @@ import UIKit
     ///
     /// - Parameter presentingViewController: The view controller from which to present the Link sheet.
     /// - Parameter email: The email address to pre-fill in the Link sheet. If `nil`, the email field will be empty.
-    /// - Parameter supportedPaymentMethodTypes: The payment method types to support in the Link sheet. Defaults to all available types.
+    /// - Parameter supportedPaymentMethodTypes: The payment method types to support in the Link sheet. If `nil`, all available types are supported.
     /// - Parameter collectName: Whether or not we should collect the user's name and attach it to the billing details.
     /// - Parameter completion: A closure that is called when the user has selected a payment method or canceled the sheet. If the user selects a payment method, the `paymentMethodPreview` will be updated accordingly.
     @_spi(STP) public func collectPaymentMethod(
         from presentingViewController: UIViewController,
         with email: String?,
-        supportedPaymentMethodTypes: [LinkPaymentMethodType] = LinkPaymentMethodType.allCases,
+        supportedPaymentMethodTypes: [LinkPaymentMethodType]? = nil,
         collectName: Bool = false,
-        completion: @escaping () -> Void
+        completion: @escaping (_ didSelectPaymentMethod: Bool) -> Void
     ) {
         var configuration = self.configuration
         configuration.defaultBillingDetails.email = email
@@ -395,12 +395,12 @@ import UIKit
                 if shouldClearSelection {
                     self?.internalPaymentOption = nil
                 }
-                completion()
+                completion(false)
                 return
             }
 
             self?.internalPaymentOption = .link(option: confirmOption)
-            completion()
+            completion(true)
         }
     }
 
@@ -1021,13 +1021,13 @@ extension LinkController: LinkFullConsentViewControllerDelegate {
     ///
     /// - Parameter presentingViewController: The view controller from which to present the Link sheet.
     /// - Parameter email: The email address to pre-fill in the Link sheet. If `nil`, the email field will be empty.
-    /// - Parameter supportedPaymentMethodTypes: The payment method types to support in the Link sheet. Defaults to all available types.
+    /// - Parameter supportedPaymentMethodTypes: The payment method types to support in the Link sheet. If `nil`, all available types are supported.
     /// - Parameter collectName: Whether or not we should collect the user's name and attach it to the billing details.
     /// - Returns: A `PaymentMethodDisplayData` if the user selected a payment method, or `nil` otherwise.
     func collectPaymentMethod(
         from presentingViewController: UIViewController,
         with email: String?,
-        supportedPaymentMethodTypes: [LinkPaymentMethodType] = LinkPaymentMethodType.allCases,
+        supportedPaymentMethodTypes: [LinkPaymentMethodType]? = nil,
         collectName: Bool = false
     ) async -> LinkController.PaymentMethodPreview? {
         return await withCheckedContinuation { continuation in
@@ -1037,9 +1037,9 @@ extension LinkController: LinkFullConsentViewControllerDelegate {
                     with: email,
                     supportedPaymentMethodTypes: supportedPaymentMethodTypes,
                     collectName: collectName
-                ) { [weak self] in
+                ) { [weak self] didSelectPaymentMethod in
                     guard let self else { return }
-                    continuation.resume(returning: self.paymentMethodPreview)
+                    continuation.resume(returning: didSelectPaymentMethod ? self.paymentMethodPreview : nil)
                 }
             }
         }