+++ CHANGELOG.md
@@ -3,6 +3,9 @@ MINOR
 
 ## X.Y.Z - changes pending release
 
+### Identity
+* [Improved] Improved StripeIdentity analytics with richer error details and screen/camera context to help debug verification flows.
+
 ### All
 * [Removed] Removed UPI support across the SDK.
 
@@ -11,6 +14,9 @@ MINOR
 
 ### Connect
 * [Changed] [#66351](https://github.com/stripe/stripe-ios/pull/6351) Theming tokens in connect embedded components is now GA.
+
+### CryptoOnramp (Alpha)
+* [Fixed] Fixed an issue (#6363) in which StripeCryptoOnramp was not properly exposed via Swift Package Manager.
 
 ## 25.11.0 2026-04-13
 ### PaymentSheet